Refactored main.py for workflows
