+++ flomps_algorithm/output/flam_20241028_131830.txt
@@ -1,0 +1,5 @@
+         time_stamp  satellite_count satellite_name  aggregator_flag federatedlearning_adjacencymatrix
+2024-03-01 12:00:00                3    Satellite 2             True [[0, 0, 0], [0, 0, 1], [0, 1, 0]]
+2024-03-01 12:30:00                3    Satellite 1             True [[0, 0, 1], [0, 0, 0], [1, 0, 0]]
+2024-03-01 13:00:00                3    Satellite 3             True [[0, 0, 0], [0, 0, 1], [0, 1, 0]]
+2024-03-01 13:30:00                3    Satellite 1             True [[0, 0, 1], [0, 0, 0], [1, 0, 0]]
