+++ sat_sim/output/sat_sim_20241028_130458.txt
@@ -1,0 +1,23 @@
+Number of satellites: 3
+Satellite 1, Satellite 2, Satellite 3
+
+Time: 2024-08-12 12:00:00
+0 0 1
+0 0 0
+1 0 0
+
+Time: 2024-08-12 12:30:00
+0 0 0
+0 0 0
+0 0 0
+
+Time: 2024-08-12 13:00:00
+0 0 1
+0 0 0
+1 0 0
+
+Time: 2024-08-12 13:30:00
+0 0 0
+0 0 0
+0 0 0
+
